fix(scripts): refuse a non-local database in every destructive script
A repo-wide audit, run because one seed script turned out able to delete
a tenant it never verified, found the same shape in the main seed and
worse.

drizzle/seed.ts's cleanFramework resolves its delete set from the GLOBAL
requirement catalog: requirementCategory -> requirement ->
companyRequirementStatus. requirement has no companyId, and the tenancy
on those tables runs through assessmentId -> companyAssessment.companyId,
which the function never joins. The word companyId does not appear in it.
So it deletes evidence, requirement assignments, requirement statuses,
category assignments and intake answers for EVERY company in the
database. There is no transaction, and sign_off_history references
company_requirement_status with ON DELETE no action, so on any database
carrying one real sign-off it commits the evidence and assignment
deletes and then aborts — destroying two tenanted tables and reseeding
nothing.

Its guard was NODE_ENV === "production", which is inverted in both
directions. `bun run` leaves NODE_ENV undefined, so it never fired under
the invocation the file documents. The Dockerfile sets it to
"production", so it fired only inside the image, where the script never
runs. Meanwhile docs/coolify-deployment.md listed `bun run db:seed` as a
manual production step.

The repo already had the right pattern and had not applied it where it
counted. scripts/smoke-supplier-portal.ts checks the connection itself,
including the `?host=` parameter that pg-connection-string prioritises
over the URL authority, so `postgres://u:p@localhost/db?host=prod` reads
as local while connecting to prod. Lifted to
scripts/lib/assert-local-database.ts and applied to the three scripts
that lacked it: drizzle/seed.ts, scripts/apply-migration.ts (executes
arbitrary SQL, no transaction, no journal entry) and
scripts/eval-panel.ts (picks a tenant with an unfiltered findFirst and
sends that company's compliance report to an external model). The smoke
script now imports it rather than keeping its own copy.

Connection-shaped rather than an env flag, deliberately: a flag is set
once, lands in .env beside the production DATABASE_URL, and is quietly
true for every run after. A host cannot be set and forgotten because it
is the thing being pointed at. Honest limit, stated in the helper: a
prod database fronted by ssh -L presents as localhost and passes. It
stops an accident, not a determined foot-gun.

Also corrects the deployment doc, which now says local-only and explains
that a fresh production database needs no seed at all, since framework
reference data ships as a migration.

Verified: all four scripts refuse a remote host, the ?host= bypass is
still caught, and a genuinely local URL passes.

Not fixed here, and the deeper problem: cleanFramework should not delete
tenant rows at all. Reference data wants the upsert-by-natural-key path
that seedFramework, seed-gdpr and the migration generator already use.
That is a restructure of the seed rather than a guard, and it should not
ride along with a fix whose whole point is reducing blast radius.
